pyproject.toml
[project]
name = "django-allauth-thaid"
version = "0.1.0"
description = "ThaiD OAuth2 provider for django-allauth"
readme = "README.rst"
license = "MIT"
authors = [
    { name = "Wichit Sombat", email = "wichit.s@ubu.ac.th" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-allauth>=0.63.0",
    "requests>=2.32.5",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
]

[tool.hatch.build.targets.wheel]
packages = ["allauth_thaid"]

[project.urls]
Homepage = "https://github.com/wichit2s/django-allauth-thaid"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "build>=1.4.2",
]
